+++ WebContent/WEB-INF/header.jsp
@@ -1,11 +1,9 @@
-<%@ page language="java" contentType="text/html; charset=utf-8"
-    pageEncoding="utf-8"%>
 		<header >
 				<h1 > Agriotes </h1 >	
 			<nav >
 				<ul >
-					<li> <a  href = " index.jsp "  class = " index " > Accueil </a> </li >
-					<li> <a  href = " # " > PrÃ©sentation </a >
+					<li> <a  href = "accueil"  class = " index " > Accueil </a> </li >
+					<li> <a  href = " # " > Présentation </a >
 						<ul >
 							<li > <a  href = " # " > Le centre </a > </li >
 							<li > <a  href = " # " > Les EQUIPES </a > </li >
@@ -20,7 +18,7 @@
 						</ul >
 					</li >
 					<li > <a  href = " " > S'inscrire </a> </li >
-					<li > <a  href = " Connexion.jsp " > Se connecter </a > </li >
+					<li > <a  href = " connexion.jsp " > Se connecter </a > </li >
 				</ul >
 			</nav >
 		</header >
