+++ WebContent/WEB-INF/headerFormateur.jsp
@@ -2,7 +2,7 @@
 	<h1>Agriotes</h1>
 	<nav>
 		<ul>
-			<li><a href=" index.jsp " class=" index "> Accueil </a></li>
+			<li><a href="accueil" class=" index "> Accueil </a></li>
 			<li><a href=" # "> GESTION </a>
 				<ul>
 					<li><a href=" # "> Ressources </a></li>
@@ -15,7 +15,7 @@
 				<ul>
 					<li><a href=" # "> QCM </a></li>
 					<li><a href="liste-evaluations"> Evaluer </a></li>
-					<li><a href=" # "> RÃ©sultat </a></li>
+					<li><a href=" # "> Résultat </a></li>
 				</ul></li>
 			<li><a href=" # "> PROJET </a>
 				<ul>
@@ -28,7 +28,7 @@
 					<li><a href=" # "> Afficher </a></li>
 					<li><a href=" # "> Modifier </a>
 					<li><a href=" # "> Changer mot de passe </a></li>
-					<li><a href=" # "> Se dÃ©connecter </a></li>
+					<li><a href=" # "> Se déconnecter </a></li>
 				</ul></li>
 		</ul>
   </nav>
